port of commit 3b7011d017a
Author: Paul Wouters <paul@xelerance.com>
Date:   Mon Mar 30 09:11:06 2009 -0400

Fix for a deadly 1 packet crasher in the DPD processing code. This bug
is severe and has been assigned CVE-2009-0790. All versions of Openswan,
Superfreeswan and Strongswan to date suffered from this bug. If the
icookie/rcookie did not match for a DPD R_U_THERE or R_U_THERE_ACK, the
pluto daemon would crash (and restart). This can be abused in a denial of
service attack.