There are a number of bugs with the SA refcount usage in klips.  This is causing SA's to hang around and their memory not to be freed.  For OCF systems this can run HW out of resources.
Most of it stems from pluto's assumption that deleting an SA that is part of
a group will clean up the whole group,  but the klips SA group removal code
is not complete.

Firstly,  ipsec_sa_rm and ipsec_untern need to be called on each SA in the
group to reduce the reference counts.

Further to this, on a group SA,  pluto will free only one,  for example the
most comon would be the ESP SA.  But there are IPIP and potentially others.
Depending on the direction the order of the SA's is changed.

Because the current grouping code was a simple linked list,  if the ESP SA
was not at the head of the list (due to ordering) the others would be missed.

To fix this we now have a doubly linked list so we can easily find the head
of the SA group when performing group operations.

There were also a few instances of underflow refcounting that are fixed by
this patch.

This patch implements the minimal fixes needed at this time to cleanup the
klips refcount usage.  There will be a followup patch that contains the
tracking code used to find these problems and clean it up :-)