Fix boundary conditions within ipsec_sa_recycle
There are some problems in the ipsec_sa_recycle() function when the
refFreeListCont swaps from one subtable to the next or when the end of the
whole refTable is reached. refFreeListHead should also only be set to
IPSEC_SAREF_FIRST, when a free "slot" can be found, because otherwise it
points to an illegal entry. For more details, see the attached patch.

Martin Schiller <mschiller@tdt.de>