* Workaround for Android Ice Cream Sandwich ipsec-tools 0.8.0 bug
ipsec-tools 0.8.0 mistakenly sets some NAT-OA fields that are defined
in RFC1374 as "always zero". We define these as "ft_mbz" (Must Be Zero)

This workaround changes the type to "ft_nat" (Natural number) and
then ignores it.

What we really need is the "ft_mbz" case to log and zeroise, but I
could not get the pointer magic working.