* Pass traffic selectors to the kernel in Transport Mode
2. Traffic selectors in transport mode (both ikev1/ikev2) (redhat
bz#831669): Openswan does not pass traffic selectors information to
kernel during setup of SAs when a connection is configured in transport
mode. This might lead to situation where esp packets not matching to
existing traffic selectors can pass through kernel when the SA is in
transport mode. The attached patch (openswan-831676.patch) addresses
this issue and now Openswan passes traffic selectors information to
kernel when SAs are setup in transport mode.