* WIP testing: Added scripts for network reconfiguration based on test host
during kickstart %post, we add the osw-bindmount.service for systemd. This
file is created inline from the kickstart file, as we do not yet have the
/testing directory mounted while we are installing.

Once the VM boots, it is passed a umid=hostname kernel option, which this
service uses to bind-mount /etc/sysconfig/network* before the network service
is started.

This allows us to use one disk image for all VMs we fire up. We use a disk
image because it is created on the fly with virt-create, and not all VM
technologies can do hostfs like uml can. (containers/namespaces can, but
those do not support cgroups/namespaces for XFRM/NETKEY)