* testing: remove baseconfig/testing/bin/rs
The "rs" script was to "restart" ipsec within the uml when it has
been recompiled on the host. This saves a reboot of the uml, but
it uml specific (and caused problems on fedora's rootfs where
the bin/sbin directories are 555, not 755)

Only old master probably knew and used this shell script.