* testing: change from transient to persistent networking in libvirt
I had not realised virsh net-create only creates a "temporary" network.
We have to use net-define to make it permanent. The xml files don't
define whether to start it now or on boot, so we have to issue virsh
commands for that as well.

Also added the 192.9.2.0/24 and 192.9.4.0/24 networks (eth2 on the umls)