* XAUTH/X509 locking issues reworked
There are several problems with the locking code being partially
bypassed, and the crl fetch function being compiled into an empty
function. See bugs #1390, #1391, #1392.

This patch changes the HAVE_THREADS define, which was misleading
because threads are dependant upon in all cases for the crypto
helpers. Instead, it really referred to CRL/LDAP fetching, and
XAUTHPAM code. So most of the defines got renamed to signify that.
This exposed some problems with the HAVE_THREADS define not being
exposed in all factored out libraries in lib/

The check_crls() function now never compiles into an empty stub. The
function is either defined when CRL fetching is enabled, or it is not
defined. It also had to be moved for this.

The CRL/LDAP fetching code was refactored into segments and ended up
residing in programs/pluto/x509*, programs/pluto/fetch.c and
lib/lib*swan/x509*. However, some of this code used a global
variable inside programs/pluto/x509.c, so this refactoring never
worked (and only compiled because the HAVE_THREADS define was lost
when compiling the moved code into lib/lib*swan/x509*c resulting in
empty stubs or skipped segments)

Note that HAVE_OCSP is never set (it was removed from Makefile.inc
a long time ago), so any code using this ifdef is never used. I will
remove that code (it also should be done using NSS)