* _plutorun: iIPSEC_CONFDIR check for /etc/ipsec.d/ override was broken
It always passed --ipsecdir /etc/ipsec.d, even when no IPSEC_CONFDDIR
was passed. Because it always set this option, it would not look for
ipsec.d within IPSEC_CONFS as prefix. This was causing the test cases
to fail as the NSS files are in ipsec.d, and it looked in /etc/ipsec.d
instead of in /tmp/TESTNAME/ipsec.d

(note also the double D in IPSEC_CONFDDIR - a little misleading, but I
left it unchanged as to not to break other people's setup