* NAT-T: Fix all broken logging of NAT-T methods / vendor ids
We now always display the method name (which is not the VID name)
and never a number. It used to try and lookup the vendorid name
against the vendormethod, and hence it failed to display text.

I changed things to use a proper enum_name. However, due to the LELEM
use of the method stored in st_nat_traversal, displaying the proper
method there was rather convoluted. But I refrained from splitting
st_nat_traversal into two variables (one for current received prefered
natt vid and one for the natt status detection (us, them or both NATed)