* change ft_mbz to ft_zig
This was a pending change after talking to both dhr and Avesh.

We now continue and ignore all receiving bits that should be zero
but are not in in_struct()

In out_struct(), code was already in place to zeroize what we believe
must be zero for our outgoing packets.

This makes us liberal in what to receive, and strict in what we sent.
It also removes having two different "zeroish" fields of ft_mbz/ft_zig,
as our payload definitions are used symmetrically, so the distinction
was a little weird.