* scripts: phased out /var/run/pluto/ipsec.info
This information is now obtained in the scripts directly. A lot of
this usage will probably get phased out when _plutorun/_plutoload will
be removed. This still needs testing when used with point-to-point
interfaces.

It seems this information is mostly used for ipsec tncfg to bind
the physical interface to the virtual interface, and to give
the default source ip and nexthop to addconn, which will feed this
into pluto. The new pluto should probably just not care, and pick
whatever is the default IP and nexthop, but that needs some more
careful thinking (with DHR)