* _include: Removed obsolete _include
It's only left call was in ipsec barf, where it was used to dump
the ipsec.conf and ipsec.secrets (filtered through a key censor)

We now have readwriteconf that does that using our C parser for
ipsec.conf (include file aware), so we use that.

For the ipsec.secrets file, we don't have an alternative, as we don't
currently have a readwritesecretconf type test that's run in "make check"
like we do for the readwriteconf tests. So for now we use "cat". So
we will miss the included secrets for now. But this is only a debug tool,
so not very essential.