* systemd: put working systemd service file for the pluto daemon.
This is a native systemd service file that depends on the new --config /etc/ipsec.conf
option that makes pluto read all its parameters from "config setup". This
obsoletes _plutoload and _plutorun completely.

What still needs to be finished is the replacement script for the kernel module
preparations - currently somewhat drafted in stackmanage.in