* ipsec verify has been extended with more checks
It checks ipsec.conf and ipsec.secrets for syntax errors.
Obsoleted ipsec.conf options are displayed as warning
The OE checks are back, and check FQDN/myid and IPSECKEY in the forward DNS.
The command now return the number of failed tests in text and as exit code.

Man page updated to reflect removed --host option