* XAUTH: Added xauthby=alwaysok option
Setting xauthby to alwaysok causes the XAUTH authentication to always succeed.
This is useful to supoprt clients that require XAUTH, but for which no real
XAUTH usernames/passwords are provisioned. This is valid for some certificate
based deployments of devices.

The static function do_md5_authentication() got renamed to do_file_authentication(),
because it is using the crypt() call, which supports more then just MD5.

The man page has been updated to reflect this, and also adds a note about MD5 not
being available in FIPS mode.

A separate bug has been opened for a feature to set the xauth password file name,
instead of hardcoding it to /etc/ipsec.d/passwd.