* pluto: show orientation with ipsec auto --status
When a connection is not oriented, the display of such a connection
in ipsec auto --status is 'undefined'. One side is called "left" without
any real proof. As such, one could not see the difference between a
properly oriented connection, and a unoriented connection that just
happened to look the same. This adds an entry to the output that will
state "oriented" or "unoriented", eg:

000 "redhat": 76.10.157.69[@RH-standard,+MC+XC+S=C]---76.10.157.65...66.187.233.55<vpn-rdu.redhat.com>[MS+XS+S=C]; unrouted; eroute owner: #0
000 "redhat":     oriented; myip=unset; hisip=unset;
000 "redhat":     xauth info: myxauthuser=pwouters;
000 "redhat":   ike_life: 3600s; ipsec_life: 28800s; rekey_margin: 540s; rekey_fuzz: 100%; keyingtries: 0; sha2_truncbug: yes
000 "redhat":   policy: PSK+ENCRYPT+TUNNEL+PFS+DONTREKEY+XAUTH+AGGRESSIVE+IKEv2ALLOW+SAREFTRACK+lKOD+rKOD; prio: 32,32; interface: virbr0;
000 "redhat":   dpd: action:hold; delay:30; timeout:60;