* status: slight change in output of ipsec auto --status
We used to only display metric and mtu when one of these were set.
We now always display these. The prio and interface were moved on
their own line with metric and mtu. This gives us more space for
our ever increasing list of POLICY bits to be displayed.

old:
000 "redhat":   policy: PSK+ENCRYPT+TUNNEL+PFS+DONTREKEY+XAUTH+AGGRESSIVE+IKEv2ALLOW+IKE_FRAG; prio: 32,32; interface: virbr0;

new:

000 "redhat":   policy: PSK+ENCRYPT+TUNNEL+PFS+DONTREKEY+XAUTH+AGGRESSIVE+IKEv2ALLOW+IKE_FRAG;
000 "redhat":   prio: 32,32; interface: virbr0; metric:0, mtu:unset;

For OE, the DNS policies (+lKOD and +rKOD) are added to the policy line, but after the ";"
to avoid confusing thinking these are c->policy bits.