* #53:  ipsec auto --status does not show phase2 parameters
when using (unspecified) defaults

Not specifying phase2alg= leaves c->alg_info_esp NULL so the rest of the
information was being skipped, when c->alg_info_esp was only needed to
determine the pfsgroup in whack_log. Relocating the pfsgroup determination
outside of the whack_log functions will let us see the rest of the info
even if the pfsgroup is unspecified.

Signed-off-by: Paul Wouters <pwouters@redhat.com>