* testing: swan-prep was importing p12 files without starting fresh
It also caused it to prompt for a password when initiating the nss
db files. It now runs it through pexpect to create it, then fills
it in with the right p12 file.

Currently it then imports the public certs of east, west and road.
This is needed for rightcert=XXXX when there is no CA and certs
don't come in over IKE. Test cases that want to use the CA should
delete these public certs.