* testing: fixup of runkvm.py
I had introduced the skipping of lines starting with "#" a while ago
to fix text flow issues, but that caused us to not put in the markers
in the console log for # --- cut --- and # --- tuc --- and we would
end up with too much for the sanitized console.

runkvm.py also called ipsec whack shutdown, even though we do that already
in final.sh. At for non-pluto userlands it would need to be different anyway.
And for some tests (eg netkey) we want to test if the ip xfrm tables are
empty afterwards, so it is not neccessarily the last action we want to do.
So leave it up to final.sh to do the shutdown.