* testing: Make dotest.sh and runkvm.py a little more robust
I was regularly seeing "hangs" and a failure for a test case to take
control of a VM. These changes resolve the hangs for me and improves
the output for humans to figure out what's going on.

We used the serial console to reboot a VM, now we use virsh reboot
directly. When we still had something running (eg ipsec auto --up retrying
a long time) our reboot command would never arrive.

Related, we now hit return and ctrl-c when reconnecting to the
VM. Additionally, we run stty sane because the console's tty insanity upon
reconnect seems to cause what we called "flow problems" in the output.

Use setproctitle to rename the runkvm.py processes (called python) to
"swankvm". This allows us to "killall swankvm" at the start of a test
to kill any lingering python scripts from previous runs. We also kill
any remaining tcpdump processes.

Prepend the prompt (hostname@testname:) for all output to the shell
running the test, to make it easier for the human to see which of the
running hosts is generating the output while the test is running. This
does not change the output in the test OUTPUT/ directory

wrap all child.expect() calls into a try: / except: statement, so we can
just throw a human readable error, instead of a python stack trace that
scrolls off the screen, especially when sharing a screen with 'screen'.

When expect is waiting on either the login: prompt or the root prompt,
act differently based on which we actually get back. Only attempt to
login when we did not get a root prompt.

Move deletion of /tmp/pluto.log and symlink from runkvm.py to swan-prep

Also reboot "nic" for each test so it properly clears the iptables and
conntrack tables.

Reduce the timeout values for expect so failing tests fail a little quicker.
Before this it would take minutes to fail.

Clearly notify failure/success for gaining access to a VM.

When hitting return to get a shell prompt, also attempt ctrl-c