* testing: Figured out the occasional mangling of lines!!
The cause was that we were waiting in the expect loop on the prompt
to return, to then send the next line. But the prompt was defined
as "root@hostname", even though it was "[root@hostname testname]# ".

probably the [] were left out because within expect that also has
meaning so you have to protect them using \[ and \]

With matching the full prompt now, I managed to run basic-pluto-01
5 times in a row without seeing the mangling anywhere.