* IKEv1: Only mark peer as fragment capable after assembling a fragment
We used to mark a peer as fragment-capable after receiving a first
fragment. Now we wait until we have assembled a full IKE packet from
fragments.

Regardless, when we receive the vendorid we deem them fragment capable.
In theory this could be spoofed, but an attacker that can modify packets
can do a DOS anyway.