* testing: basic-pluto-02 fixup as it likely was meant to be.
Since part of the "known good output" was missing, west specifically,
I have to take a guess at what this was supposed to do. I believe it
is meant to reject the connection on east because the eastnet-westnet
conn is explicitely not loaded, and the OE conn would not match such
subnets.