* permissions: open up /var/run/pluto, close down /etc/ipsec.d
The rundir (default /var/run/pluto) is changed from 700 to 755, to
allow non-root processes to read pluto.pid (eg monitor scripts)

The ipsecddir (default /etc/ipsec.d) and its subdirectories is changed
from 755 to 700. This was already the case for some distributions
(Fedora, RHEL). This provides a little more privacy about which IPsec
tunnels are configured, which certificates are known, etc.