This is fix for libreswan bug #85.
We only add traffic selectors for transport mode. The problem is that
Tunnel mode ipsec with ipcomp is layered so that ipcomp tunnel is
protected with transport mode ipsec but in this case we shouldn't any
more add traffic selectors or we break the tunnel.
Function setup_half_ipsec_sa was modified to inform netlink_setup_sa with
add_selector boolean about need to add selectors. This prevents breaking
ipcomp in tunnel mode. Direction of sa is now passed to netlink_setup_sa
so client can be substituted with host ip so that selector works for natted
transport mode.

Signed-off-by: Tuomo Soini <tis@foobar.fi>