* security: Fix misuse of ASN1_BUF_LEN and PATH_MAX in unused function
ASN1_BUF_LEN was declared as 512 at one location and 256 at another. Moved
it to x509.h as 512 and verified its usage.

In load_host_cert() it copied a char[PATH_MAX] for ASN1_BUF_LEN length. Luckilly,
PATH_MAX is 1024 so it would always fit.

Additionally, since load_host_cert() is only used in the non-NSS code branch
which we no lnoger support, this function has been #if'defed out.