* Fix for CRL signature verification failure if first byte is a zero
ASN.1 integer values have a leading zero if they are above a certain size to
denote whether they are positive or negative values. Oddly enough, the signature
is stored as an integer value.

The CRL verification code introduced with the NSS code path used the gmp library's
bignum to convert the signature chunk_t into a bignum and back, thereby removing
the leading zero. However, this would remove more then the 1 leading zero, so if
the signature started with a 0x00, then the RSA signature of the CRL would be short
a byte and fail to verify. The CRL would be rejected.

This patch removes the conversions to bignum, and handles the leading zero by just
moving the pointer one forward, and reducing the length by 1.

Debugging was also slightly cleaned up, and errors in the CRL are now reported back
to the user if reading the CRLs was triggered by "ipsec auto --rereadall

(this is rhbz#959969)