netkey: restore port when ipsec policy is generated for nat-t
This is libreswan bug #101.

Save and restore port when mangling ip addresses for xfrm policy with
nat-t transport mode. We got wrong port from struct because we copied
host ipaddress to client ipaddress when creating ipsec policy for
transport mode with XFRM/NETKEY because host port was copied with ip
to client port. And because the port was wrong we failed
to delete the ipsec policy when transport mode IPsec SA was taken down.

Signed-off-by: Tuomo Soini <tis@foobar.fi>