pluto: fixup some NSS logging to clarify pointers vs content
Upon reading the logs, some debug statements appeared to be printing
the content of a struct (eg chunk_t) instead of just the pointer to it,
likely because with nss we don't have all the content as some of it is
guarded inside nss and we only have a pointer. for some reason, the
pointer returned by nss often contains "7f 00 00", so this was causing
some confusion as being misrepresented as content, for instance with
the shared secret content itself.

This patch clarifies that by adding "(pointer)" as it was done with
other nss debug prints.