pluto: nhelpers= does not default to -1 (lsbz#126)
If pluto was started using --config /etc/ipsec.conf, we would read the
nhelpers= line from "config setup". If that line was not specified, we
defaulted to 0. When not using --config, we defaulted to -1.

ipsecconf_default_values() now sets nhelpers to -1, so that when --config
is used without an nhelpers= line, we do the right thing.

(related: currently we cannot specify "-1")