pluto: Simplify Pluto_IsFIPS()
Change Pluto_IsFIPS() to work similarly to Pluto_IsSElinux() by returning
an int with 0,1,2 (disabled,enabled,error).

Don't log when Pluto_IsFIPS() is called in the regular (0,1) case, let
the caller log. As Pluto_IsFIPS() was called twice, we ended up with
two log messages about FIPS being enabled.

We still call it twice, as I am not sure if FIPS would allow me to store
the results of a "fips check" in a local variable.