_stackmanager: add support back for fragicmp= and fix man page
_stackmanager did not support the fragicmp= option that _realsetup
did. This commit fixes that.

Although the man page said fragicmp defaults to yes, and the parser
set it to yes as well, I cannot find any openswan 2.x version that
actually defaulted it to yes. Although there was a comment in _realsetup.in
saying IPSECfragicmp = yes, the KLIPS code actually has:

	int sysctl_ipsec_icmp = 0;

(in ipsec_xmit.c in older versions, ipsec_proc.c in newer versions)

So while the parser would claim fragicmp=yes, the KLIPS option actually has
it disabled, as can be seen from performing a modprobe ipsec, and running
cat /proc/sys/net/ipsec/icmp

I changed the parser to set it to "no" per default, and _stackmanager will
only set the proc value to 1 when it sees fragicmp=yes (the non-default)

I also updated the man page.

I opted for changing the man page instead of the code, because as far as I
tracked it down (to openswan 2.3.1) it was already disabled in KLIPS and
_realsetup.in.