KLIPS: change default for hidetos (quality of service) to yes
The ToS field was supposed to be hidden per default using the hidetos=yes
default in "config setup". This was read by _realsetup to set the ipsec.ko
option via /proc/sys/net/ipsec/tos. At least since openswan 2.0.0, and
probably earlier, this was no longer being set.

The default upon loading the KLIPS kernel module is now to enable it.
A separate commit will re-add support to _stackmanager to disable this
when the user specifies hidetos=no in "config setup"