NETKEY: rhbz#742126 Use protocol and ports for priority
(fixes most of the need for lsbz#83 as well)

For XFRM/NETKEY we calculate a priority and match on longest prefix
first. This match did not take into consideration ports or protocols,
so a two conns with the only difference of leftprotoport=17/1701 would
not get priority. This caused passthrough routines that were only more
narrowed by protoport to not get priority. With the same priority as
the routed tunnel, order of loading would determine which one was picked.