testing: add swandefault network for the swanbase installer
This is to avoid the network "default" which used to exist on older
fedora's but has been removed from newer ones. By using "swandefault"
we stay out of potential conflicts. After the installer is done, the
only interfaces on swandefault is the VM "nic" with eth3. It is used
to provide internet to the VMs for updates/installs.