pluto: Remove HAVE_OCF userland, cleanup makefiles
As the HAVE_OCF code only refers to userland, and was only supported
in the non-NSS builds, we actually no longer support it. As this was
really only useful on really old embedded systems, it is not a great
loss. This also allows us to remove libaes, libdes and liblswcrypto
related code.

Removed all OPENSSL defines as they only related to HAVE_OCF
Further cleaned out some remnants in makefiles from TAPROOM and POLICYDIR.