Remove suppor for USE_DMALLOC=
Use LEAK_DETECTIVE or run with valgrind.

dmalloc is not available on fedora, and the source fails to compile and
is full of warnings.  It seems dmalloc is historic.

Note that the dmalloc initialisation code was also mistakenly put in
the function main_inI1_outR1() instead of main() - clearly no one has
used this code.