pluto: turn -DLEAK_DETECTIVE into runtime --leak_detective option.
Note we do NOT support setting an option in ipsec.conf because by
the time we parse the config file we are past the point of making
that decision.

We check immediately at the start of main() if we got passed
--leak_detective as argument, and act. When we parse the pluto options
for real later on, we just silently eat the --leak_detective option.