pluto: Added int crypto_req_keysize(int ksproto, int algo);
This will return the RFC default key size for an algorithm.
ksproto is used to ask for v1, v2 or ESP algorithms.

This function is for use in the out_sa() functions. We need this as for
some algos, according to the RFCs we MUST send a key length, and we
currently do not always do that. For example for AES, CAST, CAMELLIA
this is required. Current code only does this for ESP_AES.

The reason we don't set a missing key size on loading the conn, is that
we don't want to break current logic on the responder. So if we loaded a
conn with ike=aes we will accept aes128, aes192 and aes256, even though on
outgoing connections for that conn we will only use the default aes128.
If we would set aes128 upon loading, we would start rejecting aes192
and aes256 on incoming connections.

(code that will implement adding OAKLEY_KEY_LENGTH and KEY_LENGTH coming soon)