IKEv1: ignore incoming ISAKMP_NEXT_SAK (AKA ISAKMP_NEXT_NATD_BADDRAFTS)
We don't support SAK (group doi) and this VID is used by old cisco
devies as a NAT traversal payload. However, the method does not support
NAT-OA and we do not support that anymore. It is recommended to update
the Cisco firmware or device that sent this.

(It this turns out to be a bigger issue, we will need to re-instate a lot
 of complexity with NAT_T_WITH_NATD_BADDRAFT_VALUES)