addconn: Another fix from Wolfgang
You need to dump the whole table when searching for the gateway
route, otherwise if the destination is known, you get only the first
matching route, which is not necessarily the gateway route.

You only interested in the default routes gateway, so skip all other
routes here.
Otherwise the first route which has a gateway will wrongly taken to set
the nexthop.

The return after resolve_ppp_peer has to be removed, because this skips
all following routes. (paul: and I made it return void)