IKEv2: Fix process_informational_ikev2() for Delete payloads [Paul/Hugh]
An incomplete Child SA has the same SPIs as its Parent SA and could
accidentally be matched by the state machine to process the Informational
Exchange. If such informational message contained an IKE SA Delete
payload, it would trigger a passert in v2_delete_my_family() which
insisted to be given a Parent SA. We now switch to the parent if detected.

The RFC forbids mixing IKE and IPsec Deletes. IKE SA delete payload
SPI numbers and SPI size MUST be zero. This was not enforced.

Syntax errors in Delete payloads were ignored instead of failed.

Convoluted SPI logging simplified.