IKEv2: first initiating exchange of Original Responder is not a retransmit
st->st_msgid_lastrecv was not initialised to v2_INVALID_MSGID, so it was 0.
This resulted in the first exchange initiated by the Original Responder,
that has msgid 0, to be misinterpreted as an IKE_INIT_SA retransmit. The
packet was then dropped.

For both recent and old retransmit checks, take the v2_INVALID_MSGID value
as special - there has never been an exchange initiated by that side.