IKEv2: Don't respond to respond messages when parent SA was not found
IKEv2 exchanges are always pair wise. One request causes one respond.
The ISAKMP_FLAGS_R marks a packet as a response (not a request)

However, this check was for ISAKMP_FLAGS_I, which is wrong. The
ISAKMP_FLAGS_I flag has nothing to do with whether a message is a
request or a response. It only indicates which party initiated the
IKE (parent) SA in the original IKE_INIT_SA exchange. This is used
only to determine which SPI/KEY we need to use to match or
encrypt/decrypt the packet.

In other words: !ISAKMP_FLAGS_I != ISAKMP_FLAGS_R